Added django ModelForm validation

+++ ksdb/fundedsite.py
@@ -1,7 +1,8 @@
 # fundedsites.py
 from django.shortcuts import render_to_response
 from django.template import RequestContext
-import requests
+import simplejson
+import copy
 
 # Create your views here.
 from ksdb.models import IdSeq
@@ -9,6 +10,7 @@ from ksdb.models import fundedsite, fundedsite_pi_link, fundedsite_staff_link, f
 
 # Allow external command processing
 from django.http import JsonResponse
+from ksdb.forms import FundedsiteForm
 
 #import settings
 import logging
@@ -16,59 +18,57 @@ logger = logging.getLogger(__name__)
 
 def fundedsite_input(request):
     if request.method == 'POST':
-        title = request.POST.get('fundedsitetitle')
-        description = request.POST.get('fundedsitedesc')
-        abstract = request.POST.get('fundedsiteabstract')
-        pis = request.POST.getlist('fundedsitepi')
-        staffs = request.POST.getlist('fundedsitestaff')
-        organs = request.POST.getlist('fundedsiteorgan')
 
         fun_id = None
         message = "You have successfully added a fundedsite."
+        success = True
+        parameters = copy.copy(request.POST)
         if request.POST.get('action') == "edit":
             fun_id = int(request.POST.get('fundedsiteid'))
             message = "You have successfull edited fundedsite "+str(fun_id)+"."
+            parameters["id"] = fun_id
+            fundedsitei = fundedsite.objects.get(id=fun_id)
+            fundedsitem = FundedsiteForm(parameters or None, instance=fundedsitei)
         else:
             fun_id = IdSeq.objects.raw("select sequence_name, nextval('fundedsite_seq') from fundedsite_seq")[0].nextval
-
-        fundedsitem = fundedsite(id = fun_id, 
-                            title = title, 
-                            description = description, 
-                            abstract = abstract, 
-                            pis= ",".join(pis), 
-                            staff= ",".join(staffs), 
-                            organs= ",".join(organs), 
-                            
-                    )
-        fundedsitem.save()
-
-        #delete and save new person fundedsite associations
-        fundedsite_pi_link.objects.filter(fundedsiteid=fun_id).delete()
-        for per in pis:
-            fundedsite_pi_linkm = fundedsite_pi_link(fundedsiteid = fun_id, personid = per)
-            fundedsite_pi_linkm.save()
+            parameters["id"] = fun_id
+            fundedsitem = FundedsiteForm(parameters)
 
 
-        #delete and save new person fundedsite associations
-        fundedsite_staff_link.objects.filter(fundedsiteid=fun_id).delete()
-        for per in staffs:
-            fundedsite_staff_linkm = fundedsite_staff_link(fundedsiteid = fun_id, personid = per)
-            fundedsite_staff_linkm.save()
+        if fundedsitem.is_valid():
+            fundedsitem.save()
+
+            #delete and save new person fundedsite associations
+            pis = request.POST.getlist('pis')
+            fundedsite_pi_link.objects.filter(fundedsiteid=fun_id).delete()
+            for per in pis:
+                fundedsite_pi_linkm = fundedsite_pi_link(fundedsiteid = fun_id, personid = per)
+                fundedsite_pi_linkm.save()
 
 
-        #delete and save new person fundedsite associations
-        fundedsite_organ_link.objects.filter(fundedsiteid=fun_id).delete()
-        for org in organs:
-            fundedsite_organ_linkm = fundedsite_organ_link(fundedsiteid = fun_id, organid = org)
-            fundedsite_organ_linkm.save()
+            #delete and save new person fundedsite associations
+            staffs = request.POST.getlist('staff')
+            fundedsite_staff_link.objects.filter(fundedsiteid=fun_id).delete()
+            for per in staffs:
+                fundedsite_staff_linkm = fundedsite_staff_link(fundedsiteid = fun_id, personid = per)
+                fundedsite_staff_linkm.save()
 
-        return JsonResponse({'Success':"True",
-                             'errors':'',
+
+            #delete and save new person fundedsite associations
+            organs = request.POST.getlist('organs')
+            fundedsite_organ_link.objects.filter(fundedsiteid=fun_id).delete()
+            for org in organs:
+                fundedsite_organ_linkm = fundedsite_organ_link(fundedsiteid = fun_id, organid = org)
+                fundedsite_organ_linkm.save()
+        else:
+            message = simplejson.dumps(fundedsitem.errors)
+            success = False
+
+        return JsonResponse({'Success':success,
                                 'Message':message})
 
     personfield = [ [str(obj.id), str(obj.firstname), str(obj.lastname)] for obj in list(person.objects.all()) ]
     organfield = [ [str(obj.id), str(obj.name)] for obj in list(organ.objects.all()) ]
-    print organfield
     data = {"action" : "New" ,
             "pis" : personfield ,
             "staffs" : personfield ,
