+++ ksdb/project.py
@@ -1,7 +1,7 @@
 # protocols.py
 from django.shortcuts import render_to_response
 from django.template import RequestContext
-import requests
+import copy, simplejson
 
 # Create your views here.
 from ksdb.models import IdSeq
@@ -9,6 +9,7 @@ from ksdb.models import project, project_site_link, fundedsite
 
 # Allow external command processing
 from django.http import JsonResponse
+from ksdb.forms import ProjectForm
 
 #import settings
 import logging
@@ -16,36 +17,39 @@ logger = logging.getLogger(__name__)
 
 def project_input(request):
     if request.method == 'POST':
-        title = request.POST.get('projecttitle')
-        abbreviation = request.POST.get('projectabbr')
-        description = request.POST.get('projectdesc')
-        sitelist = request.POST.getlist('projectsites')
 
         pro_id = None
         message = "You have successfully added a project."
+        success = True
+        parameters = copy.copy(request.POST)
+
         if request.POST.get('action') == "edit":
             pro_id = int(request.POST.get('projectid'))
             message = "You have successfull edited project "+str(pro_id)+"."
+            parameters["id"] = pro_id
+            projecti = project.objects.get(id=pro_id)
+            projectm = ProjectForm(parameters or None, instance=projecti)
         else:
             pro_id = IdSeq.objects.raw("select sequence_name, nextval('project_seq') from project_seq")[0].nextval
+            parameters["id"] = pro_id
+            projectm = ProjectForm(parameters)
         
-        projectm = project(id = pro_id, 
-                            title = title, 
-                            abbreviation = abbreviation, 
-                            description = description, 
-                            sites = ",".join(sitelist)
-                    )
-        projectm.save()
-
-        #delete and save new person protocol associations
-        project_site_link.objects.filter(projectid=pro_id).delete()
-        for site in sitelist:
-            project_site_linkm = project_site_link(projectid = pro_id, fundedsiteid = site)
-            project_site_linkm.save()
-
-        return JsonResponse({'Success':"True",
-                             'errors':'',
+        if projectm.is_valid():
+            projectm.save()
+    
+            #delete and save new person protocol associations
+            sitelist = request.POST.getlist('sites')
+            project_site_link.objects.filter(projectid=pro_id).delete()
+            for site in sitelist:
+                project_site_linkm = project_site_link(projectid = pro_id, fundedsiteid = site)
+                project_site_linkm.save()
+        else:
+            message = simplejson.dumps(projectm.errors)
+            success = False
+        return JsonResponse({'Success':success,
                              'Message':message})
+
+
     sitefield = [ [str(obj.id), str(obj.title)] for obj in list(fundedsite.objects.all()) ]
     data = {"action" : "New" ,
             "sites" : sitefield ,
